Seperate renter and apartment test respectively.
